Add a button for user input
A button was added for the user to specify the number of squares in the grid. If the user enters a larger number than is allowed, it returns the default grid.

The button and h1 elements were added to the index.html file and the styles were customized in the css file.